requirements.txt
chromadb==0.5.0
langchain==0.2.1
langchain_community==0.2.1
numpy==1.26.4
ollama==0.2.0
pandas==2.2.2
python-dotenv~=1.0.0
rapidfuzz==3.9.2
requests~=2.31.0
semantic_text_splitter==0.13.1
sentence_transformers==3.0.0
tokenizers==0.19.1
PyYAML~=6.0.1
nltk~=3.8.1
scikit-learn~=1.3.2
uvicorn~=0.23.2
fastapi~=0.103.2
starlette~=0.27.0
transformers~=4.41.2
pydantic~=2.7.1